project_files/requirements.txt
python-dotenv
google-cloud-aiplatform
numpy
streamlit
pandas
numpy
matplotlib
plotly
joblib
scikit-learn
seaborn
python-dotenv
google-generativeai
lightgbm